chore: move handshake to the SocketIoConnection
